refactor: reorganize sensor imports and move sensor classes to a dedicated directory
